requirements.txt
faiss_cpu==1.7.4
transformers==4.16.2
torch==2.1